fix doc build issues
