+++ metricbeat/module/linux/memory/data.go
@@ -15,8 +15,6 @@
 // specific language governing permissions and limitations
 // under the License.
 
-// +build linux
-
 package memory
 
 import (
