+++ metricbeat/module/linux/memory/memory_test.go
@@ -15,8 +15,6 @@
 // specific language governing permissions and limitations
 // under the License.
 
-// +build linux
-
 package memory
 
 import (
